Osposobljeno dobavljanje i prikaz poseta kao i backend komponente za slanje zdravstvenih kartona i dijagnoza

+++ Backend/src/main/java/isaps/tim18/PSW_ISA_ClinicalC_2019/repository/PosetaRepository.java
@@ -13,5 +13,5 @@ public interface PosetaRepository  extends JpaRepository<Poseta, Long> {
 
     Poseta findById(int id);
 
-    List<Poseta> findByPacijent_id(int id);
+    List<Poseta> findByPacijent_id(Long id);
 }
